requirements.txt
asgiref==3.4.1
bleach==4.1.0
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.7
django-crispy-forms==1.12.0
django-environ==0.6.0
django-tinymce==3.3.0
django-widget-tweaks==1.4.8
idna==3.2
numpy==1.21.2
packaging==21.3
Pillow==8.3.2
psycopg2==2.9.1
pyparsing==3.0.6
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
webencodings==0.5.1
